refactor(testing): Move sidecar-plugin tests and consolidate sidecar-plugin setup
