+++ .github/workflows/prestocpp-linux-build-and-unit-test.yml
@@ -355,6 +355,12 @@ jobs:
           export PRESTO_SERVER_PATH="${GITHUB_WORKSPACE}/presto-native-execution/_build/release/presto_cpp/main/presto_server"
           # Find all Test*.java files and exclude those in the cudf package
           export TESTFILES=`find ./presto-native-tests/src/test -type f -name 'Test*.java' | xargs grep -L '^package com\.facebook\.presto\.nativetests\.cudf'`
+          if [[ "${{ matrix.enable-sidecar }}" == "true" ]]; then
+            export EXCLUDED_GROUPS=""
+          else
+            # Sidecar tests require sidecarEnabled=true.
+            export EXCLUDED_GROUPS="-DexcludedGroups=sidecar"
+          fi
           # Convert file paths to comma separated class names
           export TESTCLASSES=
           for test_file in $TESTFILES
@@ -371,6 +377,7 @@ jobs:
             -pl 'presto-native-tests' \
             -DstorageFormat=${{ matrix.storage-format }} \
             -DsidecarEnabled=${{ matrix.enable-sidecar }} \
+            ${EXCLUDED_GROUPS} \
             -Dtest="${TESTCLASSES}" \
             -DPRESTO_SERVER=${PRESTO_SERVER_PATH} \
             -DDATA_DIR=${RUNNER_TEMP} \
@@ -492,118 +499,6 @@ jobs:
             -pl 'presto-native-execution' \
             -DstorageFormat=${{ matrix.storage-format }} \
             -DsidecarEnabled=${{ matrix.enable-sidecar }} \
-            -Dtest="${TESTCLASSES}" \
-            -DPRESTO_SERVER=${PRESTO_SERVER_PATH} \
-            -DDATA_DIR=${RUNNER_TEMP} \
-            -Duser.timezone=America/Bahia_Banderas \
-            -T1C
-
-  prestocpp-linux-presto-sidecar-tests:
-    needs: [changes, prestocpp-linux-build-for-test]
-    runs-on: ubuntu-22.04
-    container:
-      image: prestodb/presto-native-dependency:0.297-202602271419-160459b8
-      volumes:
-        - /usr:/host_usr
-        - /opt:/host_opt
-    concurrency:
-      group: ${{ github.workflow }}-prestocpp-linux-presto-sidecar-tests-${{ github.event.pull_request.number }}
-      cancel-in-progress: true
-    env:
-      MAVEN_OPTS: -Xmx4G -XX:+ExitOnOutOfMemoryError
-      MAVEN_FAST_INSTALL: -B -V --quiet -T 1C -DskipTests -Dair.check.skip-all -Dmaven.javadoc.skip=true
-      MAVEN_TEST: -B -Dair.check.skip-all -Dmaven.javadoc.skip=true -DLogTestDurationListener.enabled=true --fail-at-end
-    permissions:
-      contents: read
-    steps:
-      - uses: actions/checkout@v6
-        if: |
-          github.event_name == 'schedule' || needs.changes.outputs.codechange == 'true'
-        with:
-          persist-credentials: false
-      - name: Fix git permissions
-        if: |
-          github.event_name == 'schedule' || needs.changes.outputs.codechange == 'true'
-        # Usually actions/checkout does this but as we run in a container
-        # it doesn't work
-        run: git config --global --add safe.directory ${GITHUB_WORKSPACE}
-
-      # We cannot use the github action to free disk space from the runner
-      # because we are in the container and not on the runner anymore.
-      - name: Free Disk Space
-        run: |
-          # Re-used from free-disk-space github action.
-          getAvailableSpace() { echo $(df -a $1 | awk 'NR > 1 {avail+=$4} END {print avail}'); }
-          # Show before
-          echo "Original available disk space: " $(getAvailableSpace)
-          # Remove DotNet.
-          rm -rf /host_usr/share/dotnet || true
-          # Remove android
-          rm -rf /host_usr/local/lib/android || true
-          # Remove CodeQL
-          rm -rf /host_opt/hostedtoolcache/CodeQL || true
-          # Show after
-          echo "New available disk space: " $(getAvailableSpace)
-
-      - name: Download artifacts
-        if: |
-          github.event_name == 'schedule' || needs.changes.outputs.codechange == 'true'
-        uses: actions/download-artifact@v4
-        with:
-          name: presto-native-build
-          path: presto-native-execution/_build/release
-
-      # Permissions are lost when uploading. Details here: https://github.com/actions/upload-artifact/issues/38
-      - name: Restore execute permissions and library path
-        if: |
-          github.event_name == 'schedule' || needs.changes.outputs.codechange == 'true'
-        run: |
-          chmod +x ${GITHUB_WORKSPACE}/presto-native-execution/_build/release/presto_cpp/main/presto_server
-          chmod +x ${GITHUB_WORKSPACE}/presto-native-execution/_build/release/velox/velox/functions/remote/server/velox_functions_remote_server_main
-          # Ensure transitive dependency libboost-iostreams is found.
-          ldconfig /usr/local/lib
-
-      - name: Install OpenJDK8
-        if: |
-          github.event_name == 'schedule' || needs.changes.outputs.codechange == 'true'
-        uses: actions/setup-java@v5
-        with:
-          distribution: temurin
-          java-version: '17'
-          cache: maven
-      - name: Download nodejs to maven cache
-        if: |
-          github.event_name == 'schedule' || needs.changes.outputs.codechange == 'true'
-        run: .github/bin/download_nodejs
-
-      - name: Maven install
-        if: |
-          github.event_name == 'schedule' || needs.changes.outputs.codechange == 'true'
-        env:
-          # Use different Maven options to install.
-          MAVEN_OPTS: -Xmx2G -XX:+ExitOnOutOfMemoryError
-        run: |
-          for i in $(seq 1 3); do ./mvnw clean install $MAVEN_FAST_INSTALL -pl 'presto-native-sidecar-plugin' -am && s=0 && break || s=$? && sleep 10; done; (exit $s)
-
-      - name: Run presto-native sidecar tests
-        if: |
-          github.event_name == 'schedule' || needs.changes.outputs.codechange == 'true'
-        run: |
-          export PRESTO_SERVER_PATH="${GITHUB_WORKSPACE}/presto-native-execution/_build/release/presto_cpp/main/presto_server"
-          export TESTFILES=`find ./presto-native-sidecar-plugin/src/test -type f -name 'Test*.java'`
-          # Convert file paths to comma separated class names
-          export TESTCLASSES=
-          for test_file in $TESTFILES
-          do
-            tmp=${test_file##*/}
-            test_class=${tmp%%\.*}
-            export TESTCLASSES="${TESTCLASSES},$test_class"
-          done
-          export TESTCLASSES=${TESTCLASSES#,}
-          echo "TESTCLASSES = $TESTCLASSES"
-          mvn test \
-            ${MAVEN_TEST} \
-            -pl 'presto-native-sidecar-plugin' \
             -Dtest="${TESTCLASSES}" \
             -DPRESTO_SERVER=${PRESTO_SERVER_PATH} \
             -DDATA_DIR=${RUNNER_TEMP} \
