+++ presto-native-tests/src/test/java/com/facebook/presto/nativetests/sidecar/TestNativeSidecarNotSetForCatalog.java
@@ -11,13 +11,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.sidecar;
+package com.facebook.presto.nativetests.sidecar;
 
 import com.facebook.presto.nativeworker.PrestoNativeQueryRunnerUtils;
 import com.facebook.presto.testing.QueryRunner;
 import com.facebook.presto.tests.AbstractTestQueryFramework;
 import org.testng.annotations.Test;
 
+@Test(groups = "sidecar")
 public class TestNativeSidecarNotSetForCatalog
         extends AbstractTestQueryFramework
 {
